Change the nav feedback note into a learn more note
